requirements.txt
absl-py==2.1.0
addict==2.4.0
basicsr==1.4.2
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
deep-translator==1.11.4
einops==0.8.0
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
fonttools==4.55.7
fsspec==2024.2.0
future==1.0.0
gfpgan==1.3.8
grpcio==1.70.0
huggingface-hub==0.27.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.3
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.6.2
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.2.1
numba==0.61.0
numpy==1.26.3
opencv-python==4.11.0.86
packaging==24.2
pillow==10.2.0
platformdirs==4.3.6
protobuf==5.29.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-image==0.25.1
scipy==1.15.1
setuptools==70.0.0
six==1.17.0
soupsieve==2.6
sympy==1.13.1
tb-nightly==2.19.0a20250128
tensorboard-data-server==0.7.2
tifffile==2025.1.10
timm==1.0.14
tokenizers==0.21.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yapf==0.43.0
